requirements.txt
flask
flask_cors
flask_expects_json
python_dotenv